Move duplicated code for projection to MFEM GridFunctions into MFEMTransferProjector
