+++ framework/include/mfem/transfers/MultiApplibMeshToMFEMGeneralFieldTransfer.h
@@ -26,7 +26,7 @@ class MooseMesh;
 // but MFEMMesh may differ in each subapp
 // */
 
-class MultiApplibMeshToMFEMGeneralFieldTransfer : public MultiAppTransfer
+class MultiApplibMeshToMFEMGeneralFieldTransfer : public MultiAppTransfer, public MFEMTransferProjector
 {
 public:
   static InputParameters validParams();
@@ -40,9 +40,7 @@ public:
 protected:
   mfem::FindPointsGSLIB _mfem_interpolator;
   std::vector<VariableName> _from_var_names;
-  std::vector<AuxVariableName> _to_var_names;
-  void extractOutgoingPoints(mfem::ParFiniteElementSpace & to_fespace, mfem::Vector & vxyz, mfem::Ordering::Type & point_ordering); 
-  void transfer(MFEMProblem & to_problem, FEProblemBase & from_problem);
+  std::vector<AuxVariableName> _to_var_names;  void transfer(MFEMProblem & to_problem, FEProblemBase & from_problem);
   void setMFEMGridFunctionValuesFromlibMesh(const unsigned int var_index, MFEMProblem & to_problem);
   void checkSiblingsTransferSupported() const override;
 };
