+++ framework/src/mfem/transfers/MultiApplibMeshToMFEMGeneralFieldTransfer.C
@@ -68,36 +68,6 @@ MultiApplibMeshToMFEMGeneralFieldTransfer::MultiApplibMeshToMFEMGeneralFieldTran
 }
 
 void
-MultiApplibMeshToMFEMGeneralFieldTransfer::extractOutgoingPoints(mfem::ParFiniteElementSpace & to_pfespace, mfem::Vector & vxyz, mfem::Ordering::Type & point_ordering)
-{
-  const int NE = to_pfespace.GetParMesh()->GetNE();
-  const int nsp = to_pfespace.GetTypicalFE()->GetNodes().GetNPoints();
-  const int dim = to_pfespace.GetParMesh()->Dimension();
-
-  vxyz.SetSize(nsp*NE*dim);
-  for (int i = 0; i < NE; i++)
-  {
-      const mfem::FiniteElement *fe = to_pfespace.GetFE(i);
-      const mfem::IntegrationRule ir = fe->GetNodes();
-      mfem::ElementTransformation *et = to_pfespace.GetElementTransformation(i);
-
-      mfem::DenseMatrix pos;
-      et->Transform(ir, pos);
-      mfem::Vector rowx(vxyz.GetData() + i*nsp, nsp),
-            rowy(vxyz.GetData() + i*nsp + NE*nsp, nsp),
-            rowz;
-      if (dim == 3)
-      {
-        rowz.SetDataAndSize(vxyz.GetData() + i*nsp + 2*NE*nsp, nsp);
-      }
-      pos.GetRow(0, rowx);
-      pos.GetRow(1, rowy);
-      if (dim == 3) { pos.GetRow(2, rowz); }
-  }
-  point_ordering = mfem::Ordering::Type::byNODES;
-}
-
-void
 MultiApplibMeshToMFEMGeneralFieldTransfer::transfer(MFEMProblem & to_problem, FEProblemBase & from_problem)
 {
   if (numToVar() != numFromVar())
@@ -115,7 +85,7 @@ MultiApplibMeshToMFEMGeneralFieldTransfer::setMFEMGridFunctionValuesFromlibMesh(
   mfem::ParFiniteElementSpace & to_pfespace = *to_gf.ParFESpace();
   mfem::Vector vxyz;
   mfem::Ordering::Type point_ordering;    
-  extractOutgoingPoints(to_pfespace, vxyz, point_ordering);  
+  extractProjectionPoints(to_pfespace, vxyz, point_ordering);  
   const int NE = to_pfespace.GetParMesh()->GetNE();
   const int nsp = to_pfespace.GetTypicalFE()->GetNodes().GetNPoints();
   const int dim = to_pfespace.GetParMesh()->Dimension();
@@ -255,26 +225,10 @@ MultiApplibMeshToMFEMGeneralFieldTransfer::setMFEMGridFunctionValuesFromlibMesh(
       interp_vals[i] = val;
     }
   }
+
   // Project DoFs to MFEM GridFunction
-  mfem::Array<int> vdofs;
-  mfem::Vector vals;
-  mfem::Vector elem_dof_vals(nsp*to_gf_ncomp);
-  for (int i = 0; i < NE; i++)
-  {
-    to_pfespace.GetElementVDofs(i, vdofs);
-    vals.SetSize(vdofs.Size());
-    for (int j = 0; j < nsp; j++)
-    {
-      for (int d = 0; d < to_gf_ncomp; d++)
-      {
-        // Arrange values byNodes
-        int idx = to_pfespace.GetOrdering() == mfem::Ordering::Type::byNODES ?
-                  i*nsp*dim + d + j*dim : d*nsp*NE + i*nsp + j;
-        elem_dof_vals(j + d*nsp) = interp_vals(idx);
-      }
-    }
-    to_gf.SetSubVector(vdofs, elem_dof_vals);
-  }  
+  mfem::Ordering::Type libmesh_interp_ordering(mfem::Ordering::Type::byNODES);
+  projectValues(interp_vals, libmesh_interp_ordering, to_gf);
 }
 
 void
