+++ test/tests/mfem/kernels/tests
@@ -167,6 +167,7 @@
     max_parallel = 1 # schemadiff with multiple ranks
     recover = false
     restep = false
+    abs_zero = 1e-8
   []
   [MFEMLinearElasticity]
     type = XMLDiff
